add CSP selector, optimize code to use shared component and eliminate some page warnings
